Add Reusable functions in utils.py
